Masalah pada saat MainActivity.kt dan FoodEntity.kt diubah value nya dari double ke Int

+++ app/src/main/java/com/shalfa/marketsupplies/FoodEntity.kt
@@ -7,6 +7,6 @@ import androidx.room.PrimaryKey
 data class FoodEntity (
     @PrimaryKey(autoGenerate = true) val id: Int = 0,
     val namaMakanan: String,
-    val beratMakanan: Double,
+    val beratMakanan: Int,
     val jumlahStokMakanan: Int
 )
